Reorganize frontend src and adjsut imports
