+++ frontend/src/project/ProjectManagerToolbar.svelte
@@ -2,7 +2,7 @@
   import { settings, selectedProject } from "../stores.js";
   import ProjectManagerAddBinaryToProject from "./ProjectManagerAddBinaryToProject.svelte";
   import ProjectManagerAddScriptToProject from "./ProjectManagerAddScriptToProject.svelte";
-  import Toolbar from "../Toolbar.svelte";
+  import Toolbar from "../utils/Toolbar.svelte";
 
   export let focus, openProject, showProjectManager, forceRefreshProject;
   let toolbarButtons;
