+++ frontend/src/utils/SearchBar.svelte
