+++ frontend/src/utils/SerializerInputForm.svelte
@@ -92,10 +92,10 @@
 
 <script>
   import Checkbox from "./Checkbox.svelte";
-  import { calculator, splitAndCapitalize } from "./helpers";
+  import { calculator, splitAndCapitalize } from "../helpers";
   import Icon from "./Icon.svelte";
   import FileBrowser from "./FileBrowser.svelte";
-  import Button from "./utils/Button.svelte";
+  import Button from "./Button.svelte";
   export let node, element;
   let unionTypeSelect,
     _element,
