+++ frontend/src/views/BinaryModifyView.svelte
@@ -76,9 +76,17 @@
 </style>
 
 <script>
-  import { buf2hex, chunkList, calculator, hexToByteArray } from "../helpers.js";
-  import { selected, selectedResource as _selectedResource } from "../stores.js";
-  import Button from "./utils/Button.svelte";
+  import {
+    buf2hex,
+    chunkList,
+    calculator,
+    hexToByteArray,
+  } from "../helpers.js";
+  import {
+    selected,
+    selectedResource as _selectedResource,
+  } from "../stores.js";
+  import Button from "../utils/Button.svelte";
   const selectedResource = $_selectedResource;
 
   export let modifierView, dataLenPromise, resourceNodeDataMap;
