+++ frontend/src/views/CarveView.svelte
@@ -64,7 +64,7 @@
 <script>
   import { calculator } from "../helpers.js";
   import { selected, selectedResource } from "../stores.js";
-  import Button from "./utils/Button.svelte";
+  import Button from "../utils/Button.svelte";
 
   export let modifierView, resourceNodeDataMap, dataLenPromise;
   let startInput, endInput, dataLength, errorMessage;
