+++ frontend/src/views/CommentView.svelte
@@ -58,9 +58,9 @@
 </style>
 
 <script>
-  import { calculator } from "./helpers.js";
-  import { selected, selectedResource } from "./stores.js";
-  import Button from "./utils/Button.svelte";
+  import { calculator } from "../helpers.js";
+  import { selected, selectedResource } from "../stores.js";
+  import Button from "../utils/Button.svelte";
 
   export let modifierView, resourceNodeDataMap, dataLenPromise;
   let comment, startInput, endInput, dataLength, errorMessage;
