+++ frontend/src/views/EntropyView.svelte
@@ -25,7 +25,7 @@
 </style>
 
 <script>
-  import LoadingTextVertical from "./LoadingTextVertical.svelte";
+  import LoadingTextVertical from "../utils/LoadingTextVertical.svelte";
 
   import { hexToByteArray } from "../helpers.js";
   import { selectedResource, settings } from "../stores.js";
